Added Object structures and GitIgnore
Added a simple outline for SpecFilter and SpecImage (feel free to change
these!). I was working on, but did not finish, SpecImg::getRGB(). It's a
pretty complicated problem to do "correctly" (using all bands).
SpecImage::LoadFromFile expects the name of a FOLDER, which it will look
in for the TIFs.
The GitIgnore ignores TIF and PNG images. So feel free to add image
folders to the project, just make sure you clean them up later.