Watershed start
Currently, Watershed runs on the results of christianTest. img is the
image we're going to be using for Watershed. It's a global Mat -- assign
it from anywhere before you run watershed.
You will get a popup with some sliders. Make this have minimum gaps with
the sliders, then press space. Watershed will run and markers will be
applied to your color image.
Variables above Morphology_Operations are used in both metthods.

NOTE: Ideally this will be moved out of main.